Deleted htaccess
